Changed HUD to allow for floats for hudPos, fixed scaling for differing screen sizes, and changed instrument-pos parameter to use standard model/block coords rather than texture coords.

+++ main/java/minecrafttransportsimulator/dataclasses/PackMultipartObject.java
@@ -88,7 +88,7 @@ public class PackMultipartObject{
     public class PackInstrument{
         public float[] pos;
         public float[] rot;
-        public int[] hudpos;
+        public float[] hudpos;
         public float scale;
         public float hudScale;
         public byte optionalEngineNumber;
